refactor(components): [table] create-popper
